Test coverage, also split to a src/ layout
The src/ layout was the only way I found to make test coverage
work under tox.
