+++ requirements/production.txt
@@ -1,11 +1,11 @@
 #
-# This file is autogenerated by pip-compile
+# This file is autogenerated by pip-compile with python 3.6
 # To update, run:
 #
 #    ./requirements/update.sh
 #
 bids-validator==1.6.0
-    # via neurospin_to_bids (setup.py)
+    # via neurospin-to-bids (setup.py)
 certifi==2020.12.5
     # via requests
 chardet==4.0.0
@@ -26,10 +26,10 @@ isodate==0.6.0
     # via rdflib
 lxml==4.6.2
     # via prov
-mne-bids==0.6
-    # via neurospin_to_bids (setup.py)
 mne==0.22.0
     # via mne-bids
+mne-bids==0.6
+    # via neurospin-to-bids (setup.py)
 networkx==2.5
     # via
     #   nipype
@@ -54,13 +54,13 @@ packaging==20.9
     #   nibabel
     #   nipype
 pandas==1.1.5
-    # via neurospin_to_bids (setup.py)
+    # via neurospin-to-bids (setup.py)
 prov==2.0.0
     # via nipype
 pydeface==2.0.0
-    # via neurospin_to_bids (setup.py)
+    # via neurospin-to-bids (setup.py)
 pydicom==2.1.2
-    # via neurospin_to_bids (setup.py)
+    # via neurospin-to-bids (setup.py)
 pydot==1.4.2
     # via nipype
 pyparsing==2.4.7
@@ -76,7 +76,7 @@ python-dateutil==2.8.1
 pytz==2021.1
     # via pandas
 pyyaml==5.4.1
-    # via neurospin_to_bids (setup.py)
+    # via neurospin-to-bids (setup.py)
 rdflib==5.0.0
     # via
     #   nipype
@@ -102,4 +102,4 @@ urllib3==1.26.3
 
 # The following packages are considered to be unsafe in a requirements file:
 setuptools==58.1.0
-    # via neurospin_to_bids (setup.py)
+    # via neurospin-to-bids (setup.py)
