+++ requirements/test.txt
@@ -1,11 +1,13 @@
 #
-# This file is autogenerated by pip-compile
+# This file is autogenerated by pip-compile with python 3.6
 # To update, run:
 #
 #    ./requirements/update.sh
 #
 attrs==20.3.0
     # via pytest
+coverage==6.2
+    # via pytest-cov
 importlib-metadata==3.7.2
     # via
     #   pluggy
@@ -25,9 +27,15 @@ pyparsing==2.4.7
     #   -c requirements/production.txt
     #   packaging
 pytest==6.2.2
+    # via
+    #   -r requirements/test.in
+    #   pytest-cov
+pytest-cov==3.0.0
     # via -r requirements/test.in
 toml==0.10.2
     # via pytest
+tomli==1.2.3
+    # via coverage
 typing-extensions==3.7.4.3
     # via importlib-metadata
 zipp==3.4.1
