+++ setup.cfg
@@ -15,7 +15,9 @@ classifiers =
     Topic :: Scientific/Engineering :: Medical Science Apps.
 
 [options]
-packages = neurospin_to_bids
+packages = find:
+package_dir=
+    =src
 python_requires = ~=3.6
 install_requires =
     bids-validator
@@ -26,11 +28,15 @@ install_requires =
     PyYAML
     setuptools  # actually a dependency of pydeface (for pkg_resources)
 
+[options.packages.find]
+where = src
+
 [options.extras_require]
 dev =
     pre-commit
     pip-tools
     pytest
+    pytest-cov
     tox
 
 [options.entry_points]
@@ -47,8 +53,36 @@ neurospin_to_bids =
 
 [flake8]
 ignore =
-    W503
-    E501
+    # E203 (whitespace before ':') has false positive for array slicings
+    E203,
+    # these are on the default ignore list
+    E121, E126, E226, E133, E203, E241,
+    # We want line-break *before* the operator (new PEP8 style similar to math)
+    W503,
+    # Gives false positives when a name contains an uppercase acronym
+    N802
+
+[coverage:run]
+source = neurospin_to_bids
+branch = True
+
+[coverage:report]
+# Regexes for lines to exclude from consideration
+exclude_lines =
+    # Have to re-enable the standard pragma
+    pragma: no cover
+
+    # Don't complain about missing debug-only code:
+    def __repr__
+    if self\.debug
+
+    # Don't complain if tests don't hit defensive assertion code:
+    raise AssertionError
+    raise NotImplementedError
+
+    # Don't complain if non-runnable code isn't run:
+    if 0:
+    if __name__ == .__main__.:
 
 [codespell]
 skip = .tox
