+++ src/neurospin_to_bids/__main__.py
