+++ src/neurospin_to_bids/acquisition_db.py
