+++ src/neurospin_to_bids/exp_info.py
