+++ src/neurospin_to_bids/template_deface/facemask.nii.gz
