+++ src/neurospin_to_bids/utils.py
