+++ tox.ini
@@ -4,20 +4,27 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py36, codestyle
+envlist = py36, cov, codestyle
 
 [testenv]
-deps = pip-tools
+deps =
+    pip-tools
+    -crequirements/production.txt
+    -crequirements/test.txt
 # 'python -m piptools sync' guarantees that the test environment only contains
 # pinned dependencies (it will uninstall dependencies that are not in
 # requirements/*.txt).
-install_command = python -m pip install \
-                         -c requirements/production.txt \
-                         -c requirements/test.txt \
-                         {opts} {packages}
 commands_pre =
     python -m piptools sync requirements/production.txt requirements/test.txt
 commands = pytest {posargs}
+
+[testenv:cov]
+basepython = python3.6
+commands = pytest --cov --cov-config=setup.cfg \
+                  --cov-report=term --cov-report=xml {posargs}
+deps =
+    {[testenv]deps}
+    pytest-cov
 
 [testenv:codestyle]
 # pre-commit needs to clone Git repositories over https
@@ -26,14 +33,3 @@ commands_pre =  # do not inherit the piptools command
 commands = pre-commit run --all-files
 deps =
     pre-commit
-
-[flake8]
-ignore =
-    # E203 (whitespace before ':') has false positive for array slicings
-    E203,
-    # these are on the default ignore list
-    E121, E126, E226, E133, E203, E241,
-    # We want line-break *before* the operator (new PEP8 style similar to math)
-    W503,
-    # Gives false positives when a name contains an uppercase acronym
-    N802
